Adding User, Experiment, Node and Stream objects
Currently the variables for the User, Experiment, Node and Stream
objects are created by passing a dictionary that contains the variables
each object should contain. This makes it easy to create objects
straight from the JSON data. It does mean there is a risk of creating an
object without the correct variables for the functions to run so some
error checking will have to be added in future.

For now though this means that the JSON data for each type of object can
change and the objects will be created to conform to the JSON structure.