Added metaBase base class for SensorDB objects
metaBase now is the base class for the SensorDB objects. The
initialisation and update code has been centralised in this class.

It also contains functions that are common to all objects. This includes
metadata functions.

Metadata is no longer a required value for  any of the objects but it is
automatically created as an empty dictionary on initialisation.