Adjusted metdata storage and experiment creation
Metadata is now stored as a dictionary using the metadata name as the
key and the old metadata dictionary as the value. To access the metadata
as a list like it was previously use metadata.values().

When a new experiment is created the experiment object is now created
and returned by the create experiment function.

The built in json parsing of requests is now being used where possible.
The main difference should be 'r.json' replaces 'json.loads(r.text)'.